+++ ExposurePlayback/Services/ExposureAnalytics.swift
@@ -449,3 +449,20 @@ extension ExposureAnalytics: DrmAnalyticsProvider {
         }
     }
 }
+
+
+extension ExposureAnalytics: SourceAbandonedEventProvider {
+    public func onSourcePreparationAbandoned<Tech, Source>(ofSource mediaSource: Source, byTech tech: Tech) where Tech : PlaybackTech, Source : MediaSource {
+        let data = ["Message":"SOURCE_PREPARATION_ABANDONED"]
+        let trace = Playback.Trace(timestamp: Date().millisecondsSince1970,
+                                   data: data)
+        
+        let aborted = Playback.Aborted(timestamp: Date().millisecondsSince1970)
+        
+        dispatcher?.enqueue(event: trace)
+        dispatcher?.enqueue(event: aborted)
+        dispatcher?.heartbeat(enabled: false)
+        dispatcher?.flushTrigger(enabled: false)
+        dispatcher = nil
+    }
+}