+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 #### Features
 * `EMP-11449` Improved debug potential for the *Fairplay* validation process. Added `Playback.DRM` event to trace license and cerrtificate requests.
+* `EMP-11338` `ExposureAnalytics` now adopts `SourceAbandonedEventProvider`, delivering trace events on media abandonment.
 
 #### Bug Fixes
 * `EMP-11567` Complete media locator included in `Playback.Started` and `Playback.HandshakeStarted` events.